refactor: extract tool factory utils
- src/agency_swarm/tools/tool_factory.py: re-export ToolFactory from new tool_factory_utils package

- src/agency_swarm/tools/tool_factory_utils/*: add helpers for langchain adapters, file loading, openapi import/export, schema inspection, base tool adaptation

- src/agency_swarm/integrations/fastapi_utils/*: move request model builder into its own module and consume shared schema inspector
